Fix B4: widen DMX channel from uint8_t to uint16_t end-to-end
config.h already stores DmxChannel::channel as uint16_t (DMX channels run
1-512), but Device::channel, every device constructor, and set()/get()
took uint8_t - silently wrapping any device configured on channel >= 256
to the wrong channel.

Widen to uint16_t throughout: Device::channel, getChannel(), set()/get()
(base + DmxRelay/DmxServo/Strobe overrides), and the universe/channel
constructors for DmxRelay, DmxServo, and Strobe. DmxRepeater is unaffected
(no set()/get() overrides, channel hardcoded to 0).

+++ src/device/device.h
@@ -7,7 +7,7 @@ class Device
 {
 protected:
     uint8_t universe;
-    uint8_t channel;
+    uint16_t channel;
     unsigned long lastChange = 0;
 
 public:
@@ -16,8 +16,8 @@ public:
     virtual void start() {}; // allocate resources, init
     virtual void stop() {};  // deallocate resources
     virtual void flip() {};  // call to flip the state (depending on device type)
-    virtual void set(uint8_t channel, uint8_t data) {};
-    virtual uint8_t get(uint8_t channel) { return 0; }
+    virtual void set(uint16_t channel, uint8_t data) {};
+    virtual uint8_t get(uint16_t channel) { return 0; }
     virtual bool isEnabled() { return true; }
     virtual void handle() // call it from loop() if needed
     {
@@ -37,7 +37,7 @@ public:
         lastChange = millis();
     }
     uint8_t getUniverse() { return universe; }
-    uint8_t getChannel() { return channel; }
+    uint16_t getChannel() { return channel; }
     virtual uint16_t getNumberOfChannels() = 0;
 };
